This version created a layout for the contacts, checked for permnission for access of contacts, disaplayed contacts, and contains comments on how to transition from the Contact Activity to the input page.
The code was written for Contacts layout file. It will contain a list view of contacts when permission is given. The contactAcitivity file will ask for permission to access contacts, lists the contacts one by one when permission is granted, and sends the user to the input page if the selected contact isn't already in the system(comments, not complete code for this part). There is some extra code in this version that will be deleted later.