I added an intent so that others can access the contact name, phone number, and soon to be image.
I added an intent that allows the other files in the program a way to access the name and phone number of the user selected contact. The image will eventually be added so that it can also be accessed in other files but that part needs to be researched more before it will be possible.