In this part, I added the JSON file and did a rough outline of what the writing to the file should look like.
The JSON file that wil store the inputted information after the user clicks the submit button, was created. The code to write to this JSON file is outlined but it has a lot of errors that still need to be fixed. Depending on how the contact info(image, name, phone) are stored, this code might have to be changed.